PP-3838 Move middleware into own dirs
To make things a bit clearer
