+++ common/middleware/get-gateway-account/get-gateway-account.js
@@ -6,8 +6,8 @@ const logger = require('pino')()
 const {Cache} = require('memory-cache')
 
 // local dependencies
-const {renderErrorView} = require('../response')
-const adminusersClient = require('../clients/adminusers-client')
+const {renderErrorView} = require('../../response')
+const connectorClient = require('../../clients/connector-client')
 
 // constants
 const CACHE_MAX_AGE = parseInt(process.env.CACHE_MAX_AGE || 15 * 60 * 1000) // default to 15 mins
@@ -19,26 +19,26 @@ function middleware (req, res, next) {
     logger.error(`[${req.correlationId}] Failed to retrieve gateway account external id from res.locals`)
     return renderErrorView(req, res)
   }
-  const cachedService = cache.get(`${gatewayAccountExternalId}.service`)
-  if (cachedService) {
-    res.locals.service = cachedService
+
+  const cachedGatewayAccount = cache.get(gatewayAccountExternalId)
+  if (cachedGatewayAccount) {
+    res.locals.gatewayAccount = cachedGatewayAccount
     next()
   } else {
-    adminusersClient.retrieveService(gatewayAccountExternalId,
-      req.correlationId)
-      .then(service => {
-        cache.put(`${gatewayAccountExternalId}.service`, service, CACHE_MAX_AGE)
-        res.locals.service = service
+    connectorClient.retrieveGatewayAccount(gatewayAccountExternalId, req.correlationId)
+      .then(gatewayAccount => {
+        cache.put(gatewayAccountExternalId, gatewayAccount, CACHE_MAX_AGE)
+        res.locals.gatewayAccount = gatewayAccount
         next()
       })
       .catch(() => {
-        logger.error(
-          `[${req.correlationId}] Failed to load service from adminusers: ${gatewayAccountExternalId}`)
+        logger.error(`[${req.correlationId}] Failed to load gateway account from connector: ${gatewayAccountExternalId}`)
         renderErrorView(req, res)
       })
   }
 }
 
 module.exports = {
+  CACHE_MAX_AGE,
   middleware
 }
