+++ common/middleware/get-mandate/get-mandate.js
@@ -5,8 +5,8 @@ const _ = require('lodash')
 const logger = require('pino')()
 
 // local dependencies
-const {renderErrorView} = require('../response')
-const connectorClient = require('../clients/connector-client')
+const {renderErrorView} = require('../../response')
+const connectorClient = require('../../clients/connector-client')
 
 function middleware (req, res, next) {
   const mandateExternalId = _.get(res, 'locals.mandateExternalId')
