+++ common/middleware/get-service/get-service.js
@@ -6,8 +6,8 @@ const logger = require('pino')()
 const {Cache} = require('memory-cache')
 
 // local dependencies
-const {renderErrorView} = require('../response')
-const connectorClient = require('../clients/connector-client')
+const {renderErrorView} = require('../../response')
+const adminusersClient = require('../../clients/adminusers-client')
 
 // constants
 const CACHE_MAX_AGE = parseInt(process.env.CACHE_MAX_AGE || 15 * 60 * 1000) // default to 15 mins
@@ -19,26 +19,26 @@ function middleware (req, res, next) {
     logger.error(`[${req.correlationId}] Failed to retrieve gateway account external id from res.locals`)
     return renderErrorView(req, res)
   }
-
-  const cachedGatewayAccount = cache.get(gatewayAccountExternalId)
-  if (cachedGatewayAccount) {
-    res.locals.gatewayAccount = cachedGatewayAccount
+  const cachedService = cache.get(`${gatewayAccountExternalId}.service`)
+  if (cachedService) {
+    res.locals.service = cachedService
     next()
   } else {
-    connectorClient.retrieveGatewayAccount(gatewayAccountExternalId, req.correlationId)
-      .then(gatewayAccount => {
-        cache.put(gatewayAccountExternalId, gatewayAccount, CACHE_MAX_AGE)
-        res.locals.gatewayAccount = gatewayAccount
+    adminusersClient.retrieveService(gatewayAccountExternalId,
+      req.correlationId)
+      .then(service => {
+        cache.put(`${gatewayAccountExternalId}.service`, service, CACHE_MAX_AGE)
+        res.locals.service = service
         next()
       })
       .catch(() => {
-        logger.error(`[${req.correlationId}] Failed to load gateway account from connector: ${gatewayAccountExternalId}`)
+        logger.error(
+          `[${req.correlationId}] Failed to load service from adminusers: ${gatewayAccountExternalId}`)
         renderErrorView(req, res)
       })
   }
 }
 
 module.exports = {
-  CACHE_MAX_AGE,
   middleware
 }
